Refactor imports and move stuff out from __init__.py.
